+++ src/weekly_quiz_01/Grocery.java
@@ -1,8 +1,10 @@
 package weekly_quiz_01;
 
+import java.math.BigDecimal;
+
 public class Grocery extends Product{
 
-    public Grocery(String name, int price, int weight) {
+    public Grocery(String name, BigDecimal price, BigDecimal weight) {
         super(name,price,weight);
     }
 }
