+++ src/weekly_quiz_01/LargeAppliance.java
@@ -1,7 +1,9 @@
 package weekly_quiz_01;
 
+import java.math.BigDecimal;
+
 public class LargeAppliance extends Product{
-    public LargeAppliance(String name, int price, int weight) {
+    public LargeAppliance(String name, BigDecimal price, BigDecimal weight) {
         super(name, price, weight);
     }
